HU-03 – Seguimiento de Pedidos

+++ jama/src/main/java/com/cibertec/jama/repositories/pedido/PedidoExtraRepository.java
@@ -2,7 +2,6 @@ package com.cibertec.jama.repositories.pedido;
 
 import com.cibertec.jama.entities.pedido.PedidoExtra;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.CrudRepository;
 
-public interface PedidoExtraRepository extends JpaRepository<PedidoExtra, Long> {
+public interface PedidoExtraRepository extends JpaRepository<PedidoExtra, Integer> {
 }
